requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
databricks-sql-connector==4.2.5
apscheduler==3.10.4
aiosqlite==0.19.0
structlog==23.2.0
pytz==2023.3
